refactor(snark): split mod.rs — extract protocol router and task builder
Addresses review item #7 (part 3/3). No behavior change — pure code move.

snark/mod.rs (~1109 lines) shed its two most separable concerns:
- protocol.rs — the effect/routing layer: ComputeJob/ComputeResult, the impure
  snark_compute job, and the pure SnarkProtocol router.
- builder.rs  — SNARKTaskBuilder: r1cs/witness loading + recursive-circuit
  generation + per-curve proof-task assembly.

mod.rs keeps the SNARKBehaviour facade, the value types (Input/Circuit/Field +
enums), and the crypto proving types (SNARKProof/SNARKGenerator + impls). The new
files are child modules, so they reach the value/generator types' private fields
without widening visibility; SNARKTaskBuilder and SnarkProtocol stay at their public
paths via re-exports.
