+++ crates/node/src/backend/snark/mod.rs
@@ -14,7 +14,6 @@ use rings_snark::prelude::nova::provider::ipa_pc;
 use rings_snark::prelude::nova::spartan;
 use rings_snark::prelude::nova::traits::snark::RelaxedR1CSSNARKTrait;
 use rings_snark::prelude::nova::traits::Engine;
-use rings_snark::r1cs;
 use rings_snark::snark::CompressedSNARK;
 use rings_snark::snark::ProverKey;
 use rings_snark::snark::PublicParams;
@@ -27,11 +26,6 @@ use super::types::snark::SNARKProofTask;
 use super::types::snark::SNARKTask;
 use super::types::snark::SNARKTaskMessage;
 use super::types::snark::SNARKVerifyTask;
-use crate::backend::ext::Ctx;
-use crate::backend::ext::Effect;
-use crate::backend::ext::Event;
-use crate::backend::ext::Protocol;
-use crate::backend::ext::Transition;
 use crate::error::Error;
 use crate::error::Result;
 use crate::provider::Provider;
@@ -43,6 +37,11 @@ pub const NAMESPACE: &str = "snark";
 
 #[cfg(feature = "browser")]
 pub mod browser;
+mod builder;
+mod protocol;
+
+pub use builder::SNARKTaskBuilder;
+pub use protocol::SnarkProtocol;
 
 /// Task Manageer of SNARK provider and verifier
 #[derive(Default, Clone)]
@@ -136,165 +135,10 @@ impl SNARKBehaviour {
         let manager = self.inner.clone();
         let job: crate::backend::ext::ComputeFn = Arc::new(move |input: Bytes| {
             let manager = manager.clone();
-            Box::pin(async move { snark_compute(manager, input) })
+            Box::pin(async move { protocol::snark_compute(manager, input) })
         });
         provider.register_compute(NAMESPACE, job)?;
         Ok(())
-    }
-}
-
-/// A SNARK compute request, handed to the shell via [`Effect::Compute`]. Carries the
-/// task plus enough context to route the result. Never on the wire (local only).
-#[derive(Serialize, Deserialize)]
-enum ComputeJob {
-    /// Prove `task` (prover side), then reply to `reply_to`.
-    Prove {
-        /// Task id.
-        task_id: TaskId,
-        /// DID to send the resulting verify task back to.
-        reply_to: Did,
-        /// Proof task to fold/prove (boxed: far larger than the other variant).
-        task: Box<SNARKProofTask>,
-    },
-    /// Verify `verify_task` against the locally-stored proof task (verifier side).
-    Verify {
-        /// Task id (used to look up the stored proof task).
-        task_id: TaskId,
-        /// Verify task received from the prover.
-        verify_task: SNARKVerifyTask,
-    },
-}
-
-/// The result of a [`ComputeJob`], re-injected as a self-event for the pure `step`.
-#[derive(Serialize, Deserialize)]
-enum ComputeResult {
-    /// A proof was produced; `step` sends the verify task back to `reply_to`.
-    Proved {
-        /// Task id.
-        task_id: TaskId,
-        /// DID to reply to.
-        reply_to: Did,
-        /// The produced verify task.
-        verify_task: SNARKVerifyTask,
-    },
-    /// Verification finished; the boolean is already stored in the task manager.
-    Verified {
-        /// Task id.
-        task_id: TaskId,
-    },
-}
-
-/// The impure SNARK compute job (the shell side). Runs the heavy crypto and, for
-/// verification, reads the stored proof task and records the boolean result.
-fn snark_compute(manager: Arc<SNARKTaskManager>, input: Bytes) -> Result<Bytes> {
-    let job: ComputeJob = bincode::deserialize(input.as_ref()).map_err(|_| Error::DecodeError)?;
-    let result = match job {
-        ComputeJob::Prove {
-            task_id,
-            reply_to,
-            task,
-        } => {
-            let verify_task = SNARKBehaviour::handle_snark_proof_task(task.as_ref())?;
-            ComputeResult::Proved {
-                task_id,
-                reply_to,
-                verify_task,
-            }
-        }
-        ComputeJob::Verify {
-            task_id,
-            verify_task,
-        } => {
-            let Some(task) = manager.task.get(&task_id) else {
-                return Err(Error::ExtensionError(format!(
-                    "no pending SNARK proof task for {task_id}; cannot verify"
-                )));
-            };
-            let verified = SNARKBehaviour::handle_snark_verify_task(&verify_task, task.value())?;
-            manager.verified.insert(task_id, verified);
-            ComputeResult::Verified { task_id }
-        }
-    };
-    bincode::serialize(&result)
-        .map(Bytes::from)
-        .map_err(|_| Error::EncodeError)
-}
-
-/// SNARK relay protocol: a pure router over the `snark` namespace.
-///
-/// ```text
-///   step (Ctx (), Event{from, p}) =
-///     | from = self  ∧ p = Proved(id, to, vt)  ↦ ((), [Send to (Verify id vt)])
-///     | from = self  ∧ p = Verified(id)        ↦ ((), ε)
-///     | p = SNARKProof(t)                       ↦ ((), [Compute (Prove id from t)])
-///     | p = SNARKVerify(vt)                     ↦ ((), [Compute (Verify id vt)])
-/// ```
-///
-/// Provenance splits the two payload codecs purely: `from = self` is a re-injected
-/// `ComputeResult`; any other `from` is a network [`SNARKTaskMessage`]. The heavy
-/// proving/verification is described as [`Effect::Compute`] and performed by
-/// `snark_compute` in the shell, so `step` stays pure.
-#[derive(Clone)]
-pub struct SnarkProtocol;
-
-impl Protocol for SnarkProtocol {
-    type State = ();
-
-    fn namespace(&self) -> &str {
-        NAMESPACE
-    }
-
-    fn init(&self) {}
-
-    fn step(&self, ctx: Ctx<'_, ()>, event: &Event) -> Transition<()> {
-        if event.from == ctx.did {
-            let Ok(result) = bincode::deserialize::<ComputeResult>(event.payload.as_ref()) else {
-                return Transition::pure(());
-            };
-            return match result {
-                ComputeResult::Proved {
-                    task_id,
-                    reply_to,
-                    verify_task,
-                } => {
-                    let msg = SNARKTaskMessage {
-                        task_id,
-                        task: SNARKTask::SNARKVerify(verify_task),
-                    };
-                    match bincode::serialize(&msg) {
-                        Ok(payload) => Transition::with((), vec![Effect::Send {
-                            to: reply_to,
-                            namespace: NAMESPACE.to_string(),
-                            payload: Bytes::from(payload),
-                        }]),
-                        Err(_) => Transition::pure(()),
-                    }
-                }
-                ComputeResult::Verified { .. } => Transition::pure(()),
-            };
-        }
-
-        let Ok(msg) = bincode::deserialize::<SNARKTaskMessage>(event.payload.as_ref()) else {
-            return Transition::pure(());
-        };
-        let job = match msg.task {
-            SNARKTask::SNARKProof(task) => ComputeJob::Prove {
-                task_id: msg.task_id,
-                reply_to: event.from,
-                task,
-            },
-            SNARKTask::SNARKVerify(verify_task) => ComputeJob::Verify {
-                task_id: msg.task_id,
-                verify_task,
-            },
-        };
-        match bincode::serialize(&job) {
-            Ok(input) => Transition::with((), vec![Effect::Compute {
-                namespace: NAMESPACE.to_string(),
-                input: Bytes::from(input),
-            }]),
-            Err(_) => Transition::pure(()),
-        }
     }
 }
 
@@ -463,377 +307,6 @@ impl Field {
                 value: FieldEnum::Bn256KZG(<provider::Bn256EngineKZG as Engine>::Scalar::from(v)),
             },
         }
-    }
-}
-
-/// Snark builder
-#[wasm_export]
-pub struct SNARKTaskBuilder {
-    circuit_generator: CircuitGenerator,
-}
-
-#[wasm_export]
-impl SNARKTaskBuilder {
-    /// Load r1cs sand witness from local path
-    pub async fn from_local(
-        r1cs_path: String,
-        witness_wasm_path: String,
-        field: SupportedPrimeField,
-    ) -> Result<SNARKTaskBuilder> {
-        match field {
-            SupportedPrimeField::Vesta => {
-                type F = <provider::VestaEngine as Engine>::Scalar;
-                let r1cs =
-                    r1cs::load_r1cs::<F>(r1cs::Path::Local(r1cs_path), r1cs::Format::Bin).await?;
-                let witness_calculator =
-                    r1cs::load_circom_witness_calculator(r1cs::Path::Local(witness_wasm_path))
-                        .await?;
-                let circuit_generator =
-                    circuit::WasmCircuitGenerator::<F>::new(r1cs, witness_calculator);
-                Ok(Self {
-                    circuit_generator: CircuitGenerator::Vesta(circuit_generator),
-                })
-            }
-            SupportedPrimeField::Pallas => {
-                type F = <provider::PallasEngine as Engine>::Scalar;
-                let r1cs =
-                    r1cs::load_r1cs::<F>(r1cs::Path::Local(r1cs_path), r1cs::Format::Bin).await?;
-                let witness_calculator =
-                    r1cs::load_circom_witness_calculator(r1cs::Path::Local(witness_wasm_path))
-                        .await?;
-                let circuit_generator =
-                    circuit::WasmCircuitGenerator::<F>::new(r1cs, witness_calculator);
-                Ok(Self {
-                    circuit_generator: CircuitGenerator::Pallas(circuit_generator),
-                })
-            }
-            SupportedPrimeField::Bn256KZG => {
-                type F = <provider::Bn256EngineKZG as Engine>::Scalar;
-                let r1cs =
-                    r1cs::load_r1cs::<F>(r1cs::Path::Local(r1cs_path), r1cs::Format::Bin).await?;
-                let witness_calculator =
-                    r1cs::load_circom_witness_calculator(r1cs::Path::Local(witness_wasm_path))
-                        .await?;
-                let circuit_generator =
-                    circuit::WasmCircuitGenerator::<F>::new(r1cs, witness_calculator);
-                Ok(Self {
-                    circuit_generator: CircuitGenerator::Bn256KZG(circuit_generator),
-                })
-            }
-        }
-    }
-
-    /// Load r1cs sand witness from remote url
-    pub async fn from_remote(
-        r1cs_path: String,
-        witness_wasm_path: String,
-        field: SupportedPrimeField,
-    ) -> Result<SNARKTaskBuilder> {
-        match field {
-            SupportedPrimeField::Vesta => {
-                type F = <provider::VestaEngine as Engine>::Scalar;
-                let r1cs =
-                    r1cs::load_r1cs::<F>(r1cs::Path::Remote(r1cs_path), r1cs::Format::Bin).await?;
-                let witness_calculator =
-                    r1cs::load_circom_witness_calculator(r1cs::Path::Remote(witness_wasm_path))
-                        .await?;
-                let circuit_generator =
-                    circuit::WasmCircuitGenerator::<F>::new(r1cs, witness_calculator);
-                Ok(Self {
-                    circuit_generator: CircuitGenerator::Vesta(circuit_generator),
-                })
-            }
-            SupportedPrimeField::Pallas => {
-                type F = <provider::PallasEngine as Engine>::Scalar;
-                let r1cs =
-                    r1cs::load_r1cs::<F>(r1cs::Path::Remote(r1cs_path), r1cs::Format::Bin).await?;
-                let witness_calculator =
-                    r1cs::load_circom_witness_calculator(r1cs::Path::Remote(witness_wasm_path))
-                        .await?;
-                let circuit_generator =
-                    circuit::WasmCircuitGenerator::<F>::new(r1cs, witness_calculator);
-                Ok(Self {
-                    circuit_generator: CircuitGenerator::Pallas(circuit_generator),
-                })
-            }
-            SupportedPrimeField::Bn256KZG => {
-                type F = <provider::Bn256EngineKZG as Engine>::Scalar;
-                let r1cs =
-                    r1cs::load_r1cs::<F>(r1cs::Path::Remote(r1cs_path), r1cs::Format::Bin).await?;
-                let witness_calculator =
-                    r1cs::load_circom_witness_calculator(r1cs::Path::Remote(witness_wasm_path))
-                        .await?;
-                let circuit_generator =
-                    circuit::WasmCircuitGenerator::<F>::new(r1cs, witness_calculator);
-                Ok(Self {
-                    circuit_generator: CircuitGenerator::Bn256KZG(circuit_generator),
-                })
-            }
-        }
-    }
-
-    /// generate recursive circuits
-    pub fn gen_circuits(
-        &self,
-        public_input: Input,
-        private_inputs: Vec<Input>,
-        round: usize,
-    ) -> Result<Vec<Circuit>> {
-        match &self.circuit_generator {
-            CircuitGenerator::Vesta(g) => {
-                type F = <provider::VestaEngine as Engine>::Scalar;
-
-                let input: circuit::Input<F> = public_input
-                    .into_iter()
-                    .map(|(s, v)| {
-                        (
-                            s,
-                            v.into_iter()
-                                .map(|inp| {
-                                    if let FieldEnum::Vesta(x) = inp.value {
-                                        x
-                                    } else {
-                                        panic!("Wrong curve, expect Vesta")
-                                    }
-                                })
-                                .collect(),
-                        )
-                    })
-                    .collect::<Vec<(String, Vec<F>)>>()
-                    .into();
-
-                let private_inputs: Vec<circuit::Input<F>> = private_inputs
-                    .into_iter()
-                    .map(|inp| {
-                        inp.into_iter()
-                            .map(|(s, v)| {
-                                let fields: Vec<F> = v
-                                    .into_iter()
-                                    .map(|inp| {
-                                        if let FieldEnum::Vesta(x) = inp.value {
-                                            x
-                                        } else {
-                                            panic!("Wrong curve, expect Vesta")
-                                        }
-                                    })
-                                    .collect();
-                                (s, fields)
-                            })
-                            .collect::<Vec<(String, Vec<F>)>>()
-                            .into()
-                    })
-                    .collect();
-
-                let circuits = g
-                    .gen_recursive_circuit(input, private_inputs, round, true)?
-                    .iter()
-                    .map(|c| Circuit {
-                        inner: CircuitEnum::Vesta(c.clone()),
-                    })
-                    .collect::<Vec<Circuit>>();
-                Ok(circuits)
-            }
-            CircuitGenerator::Pallas(g) => {
-                type F = <provider::PallasEngine as Engine>::Scalar;
-
-                let input: circuit::Input<F> = public_input
-                    .into_iter()
-                    .map(|(s, v)| {
-                        (
-                            s,
-                            v.into_iter()
-                                .map(|inp| {
-                                    if let FieldEnum::Pallas(x) = inp.value {
-                                        x
-                                    } else {
-                                        panic!("Wrong curve, expect pallas")
-                                    }
-                                })
-                                .collect(),
-                        )
-                    })
-                    .collect::<Vec<(String, Vec<F>)>>()
-                    .into();
-
-                let private_inputs: Vec<circuit::Input<F>> = private_inputs
-                    .into_iter()
-                    .map(|inp| {
-                        inp.into_iter()
-                            .map(|(s, v)| {
-                                let fields: Vec<F> = v
-                                    .into_iter()
-                                    .map(|inp| {
-                                        if let FieldEnum::Pallas(x) = inp.value {
-                                            x
-                                        } else {
-                                            panic!("Wrong curve, expect Vesta")
-                                        }
-                                    })
-                                    .collect();
-                                (s, fields)
-                            })
-                            .collect::<Vec<(String, Vec<F>)>>()
-                            .into()
-                    })
-                    .collect();
-
-                let circuits = g
-                    .gen_recursive_circuit(input, private_inputs, round, true)?
-                    .iter()
-                    .map(|c| Circuit {
-                        inner: CircuitEnum::Pallas(c.clone()),
-                    })
-                    .collect::<Vec<Circuit>>();
-                Ok(circuits)
-            }
-            CircuitGenerator::Bn256KZG(g) => {
-                type F = <provider::Bn256EngineKZG as Engine>::Scalar;
-
-                let input: circuit::Input<F> = public_input
-                    .into_iter()
-                    .map(|(s, v)| {
-                        (
-                            s,
-                            v.into_iter()
-                                .map(|inp| {
-                                    if let FieldEnum::Bn256KZG(x) = inp.value {
-                                        x
-                                    } else {
-                                        panic!("Wrong curve, expect bn256")
-                                    }
-                                })
-                                .collect(),
-                        )
-                    })
-                    .collect::<Vec<(String, Vec<F>)>>()
-                    .into();
-
-                let private_inputs: Vec<circuit::Input<F>> = private_inputs
-                    .into_iter()
-                    .map(|inp| {
-                        inp.into_iter()
-                            .map(|(s, v)| {
-                                let fields: Vec<F> = v
-                                    .into_iter()
-                                    .map(|inp| {
-                                        if let FieldEnum::Bn256KZG(x) = inp.value {
-                                            x
-                                        } else {
-                                            panic!("Wrong curve, expect bn256")
-                                        }
-                                    })
-                                    .collect();
-                                (s, fields)
-                            })
-                            .collect::<Vec<(String, Vec<F>)>>()
-                            .into()
-                    })
-                    .collect();
-
-                let circuits = g
-                    .gen_recursive_circuit(input, private_inputs, round, true)?
-                    .iter()
-                    .map(|c| Circuit {
-                        inner: CircuitEnum::Bn256KZG(c.clone()),
-                    })
-                    .collect::<Vec<Circuit>>();
-                Ok(circuits)
-            }
-        }
-    }
-}
-
-impl SNARKTaskBuilder {
-    /// Generate proof task
-    pub fn gen_proof_task(circuits: Vec<Circuit>) -> Result<SNARKProofTask> {
-        let task = match &circuits[0].inner {
-            CircuitEnum::Vesta(_) => {
-                type E1 = provider::VestaEngine;
-                type E2 = provider::PallasEngine;
-                type EE1 = ipa_pc::EvaluationEngine<E1>;
-                type EE2 = ipa_pc::EvaluationEngine<E2>;
-                type S1 = spartan::snark::RelaxedR1CSSNARK<E1, EE1>;
-                type S2 = spartan::snark::RelaxedR1CSSNARK<E2, EE2>;
-                let circuits: Vec<circuit::Circuit<<E1 as Engine>::Scalar>> = circuits
-                    .into_iter()
-                    .map(|circ| {
-                        if let CircuitEnum::Vesta(c) = circ.inner {
-                            c
-                        } else {
-                            panic!("Wrong curve, expect vesta")
-                        }
-                    })
-                    .collect();
-                let inputs = circuits[0].get_public_inputs();
-                let pp = SNARK::<E1, E2>::gen_pp::<S1, S2>(circuits[0].clone())?;
-                let snark = SNARK::<E1, E2>::new(&circuits[0], &pp, &inputs, &vec![
-                    <E2 as Engine>::Scalar::from(0),
-                ])?;
-
-                SNARKProofTask::VastaPallas(SNARKGenerator {
-                    pp: pp.into(),
-                    snark,
-                    circuits,
-                })
-            }
-            CircuitEnum::Pallas(_) => {
-                type E1 = provider::PallasEngine;
-                type E2 = provider::VestaEngine;
-                type EE1 = ipa_pc::EvaluationEngine<E1>;
-                type EE2 = ipa_pc::EvaluationEngine<E2>;
-                type S1 = spartan::snark::RelaxedR1CSSNARK<E1, EE1>;
-                type S2 = spartan::snark::RelaxedR1CSSNARK<E2, EE2>;
-                let circuits: Vec<circuit::Circuit<<E1 as Engine>::Scalar>> = circuits
-                    .into_iter()
-                    .map(|circ| {
-                        if let CircuitEnum::Pallas(c) = circ.inner {
-                            c
-                        } else {
-                            panic!("Wrong curve, expect vesta")
-                        }
-                    })
-                    .collect();
-                let inputs = circuits[0].get_public_inputs();
-                let pp = SNARK::<E1, E2>::gen_pp::<S1, S2>(circuits[0].clone())?;
-                let snark = SNARK::<E1, E2>::new(&circuits[0], &pp, &inputs, &vec![
-                    <E2 as Engine>::Scalar::from(0),
-                ])?;
-                SNARKProofTask::PallasVasta(SNARKGenerator {
-                    pp: pp.into(),
-                    snark,
-                    circuits,
-                })
-            }
-            CircuitEnum::Bn256KZG(_) => {
-                type E1 = provider::Bn256EngineKZG;
-                type E2 = provider::GrumpkinEngine;
-                type EE1 = hyperkzg::EvaluationEngine<E1>;
-                type EE2 = ipa_pc::EvaluationEngine<E2>;
-                type S1 = spartan::snark::RelaxedR1CSSNARK<E1, EE1>; // non-preprocessing SNARK
-                type S2 = spartan::snark::RelaxedR1CSSNARK<E2, EE2>; // non-preprocessing SNARK
-                let circuits: Vec<circuit::Circuit<<E1 as Engine>::Scalar>> = circuits
-                    .into_iter()
-                    .map(|circ| {
-                        if let CircuitEnum::Bn256KZG(c) = circ.inner {
-                            c
-                        } else {
-                            panic!("Wrong curve, expect vesta")
-                        }
-                    })
-                    .collect();
-                let inputs = circuits[0].get_public_inputs();
-                let pp = SNARK::<E1, E2>::gen_pp::<S1, S2>(circuits[0].clone())?;
-                let snark = SNARK::<E1, E2>::new(&circuits[0], &pp, &inputs, &vec![
-                    <E2 as Engine>::Scalar::from(0),
-                ])?;
-                SNARKProofTask::Bn256KZGGrumpkin(SNARKGenerator {
-                    pp: pp.into(),
-                    snark,
-                    circuits,
-                })
-            }
-        };
-        Ok(task)
     }
 }
